refactor(tuners): extract workload-feature refinement into feature_refinement module
Move EMA-based workload-feature refinement out of the orchestrator into a
dedicated WorkloadFeatureRefiner collaborator. The refiner owns the static
feature priors, pending-delta accumulation, and per-generation weight flush,
mutating metric_config.workload_features in place and reaching the scoring
engine lazily via a scorer_provider callable (deferred scorer construction).
